+++ src/server/api/lib/twilio.js
@@ -16,6 +16,7 @@ import {
   messageComponents
 } from "./message-sending";
 import { symmetricDecrypt } from "./crypto";
+import { maybeOnboardNewContact } from './onboardIncoming';
 
 const MAX_SEND_ATTEMPTS = 5;
 const MESSAGE_VALIDITY_PADDING_SECONDS = 30;
@@ -26,14 +27,6 @@ const headerValidator = () => {
   if (!!SKIP_TWILIO_VALIDATION) return (req, res, next) => next();
 
   return async (req, res, next) => {
-    /*    handleIncomingMessage({
-          From: "6467047784",
-          To:"6467047711",
-          Body: "hello world",
-          MessageSid: 123,
-          MessageStatus: true
-        })
-    */
     const { MessagingServiceSid } = req.body;
     const { authToken } = await getTwilioCredentials(MessagingServiceSid);
 
@@ -351,55 +344,6 @@ async function handleDeliveryReport(report) {
     .catch(logger.error);
 
   return insertResult;
-}
-
-
-
-async function maybeOnboardNewContact(contactNumber, userNumber) {
-  const campaignId = process.env.CAMPAIGN_ID;
-  const isExistingUser = r
-    .knex("campaign_contact")
-    .select('id')
-    .where({
-      cell: contactNumber,
-      campaign_id: campaignId
-    })
-    .limit(1)
-
-  if (! isExistingUser || (isExistingUser.length && isExistingUser.length<1)) {
-    const onboardingCampaignId = process.env.ONBOARDING_CAMPAIGN || campaignID + 1;
-    const contact = {
-      first_name: "",
-      last_name: "",
-      cell: contactNumber,
-      campaign_id: onboardingCampaignId
-    };
-    let contactId = await r
-      .knex("campaign_contact")
-      .insert(contact)
-      .return('id');
-
-    if (contactId) {
-      const onboardingText =
-        'Hi there! Thanks for sending a text to CheckUpOn.Me! If you would like to be texted by one of our volunteers, please reply to this message with the word "YES".';
-
-      let organizationId = await r
-        .knex("campaign")
-        .select("organization_id")
-        .where({campaign_id: onboardingCampaignId})
-        .first();
-
-      const replyMessage = {
-        campaign_contact_id: contactId,
-        user_number: userNumber,
-        text: onboardingText,
-        contact_number: contactNumber,
-      };
-      sendMessage(replyMessage, organizationId)
-        .then(data=>console.info(`Onboarding sent to ${contactNumber}`))
-        .catch(err=>console.error(`Unable to onboard invite ${contactNumber}: ${err}`));
-    }
-  }
 }
 
 async function handleIncomingMessage(message) {
